requirements.txt
fastapi[all]
uvicorn
sqlalchemy
python-dotenv
requests
huggingface_hub
pytest
httpx
